lib: restructure import and comments for sci.
